requirements.txt
plyer
pypresence
discord.py 
python-dotenv
asyncio
aiohttp